+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Change Log
 HEAD
 ----
 * Added support for getting and setting ``Account.oof_settings`` using the new ``OofSettings`` class.
+* Added snake_case named shortcuts to all distinguished folders on the ``Account`` model. E.g.
+  ``Account.search_folders``.
 
 
 1.10.5